Fixed migration for pg

+++ db/migrate/20150204105914_add_columns_to_listing.rb
@@ -1,6 +1,6 @@
 class AddColumnsToListing < ActiveRecord::Migration
   def change
-  	add_column :listings, :acres, :double
+  	add_column :listings, :acres, :float
   	add_column :listings, :bedrooms, :integer
   	add_column :listings, :bathrooms, :integer
   	add_column :listings, :garages, :integer
